Basic source-map. No support for macroexpand and special forms.
Works by decompiling the intermediate language (thanks @leonoel).
Nodes resulting from AST points are toposorted, so tree-seq isn't good enough
for the job.

This is an injective function, it should be surjective. Macros and special
forms introduces points in the target language, breaking surjectivity. It's not
yet clear to me if it's possible to make it surjective at all.