Fix dom/props and dom/style unmounts with their respective frame
Attributes and styles have a lifecycle like dom nodes do. If the result of a
condition is to set an attribute/style, the attribute/style should be unset when
the condition doesn’t hold anymore.

Previous impl was side-effecting the dom on DAG branch mount, but were not
tracking the DAG branch unmount, resulting in a polluted DOM.