ui4/typeahead & select: fix race
When the controlled value changes and we're not picking we ask the
server for the option label to set in the input field. In the meantime
the user might start picking though, which creates a race. Therefore
when we receive the label from the server we have to re-check whether
the user is picking.