fix: keep dom/on exceptions around
The previous implementation surfaced an exception for a single
propagation step. This creates a race where the exception might not get
seen at all.

The new implementation stabilizes on the exceptional state as well. This
will also allow to e.g. model a button with 4 states:

  idle | pending | success | failure

on top of dom/on by observing its return value.

We chose not to expose the idle state since this can be modeled in
userland. A button implementation showing this will follow.