fix: compilation with advanced optimization
This change is effectively a no-op, replacing some macros for functions
with the exact same behavior. However there is a deep bug somewhere in
either Google Closure compiler or shadow or clojurescript that results
in a broken compiled JS artifact where an optimization goes wrong and
breaks behavior.

Consider that this breaks compilation

  (defn input-notified [^Yield Y]
    (when-not (let [inp (a/fget Y input), v (let [x inp] (aget x notified?))]
                (a/set inp notified? true) v)
      ((.-notifier Y))))

While this doesn't

  (defn input-notified [^Yield Y]
    (when-not (let [inp (a/fget Y input), v (aget inp notified?)]
                (a/set inp notified? true) v)
      ((.-notifier Y))))

Yes, the no-op `let` breaks the compiled artifact. This example is
macroexpanded code, the actual changeset bypasses this issue by turning
some macros to functions and somehow pleasing the Google Closure
compiler.