fix: relieve dom/listen> backpressure
this fixes todomvc-composed observe consumer not ready error

In UIs we typically care about latest events and it's fine to drop
unprocessed events, so this change is safe.

Note: relieving in e/for-event would work too, but relieving the DOM
events is a tighter fix and allows e/for-event users to pick
backpressure strategy themselves (e.g. buffer).