fix clj class type hints debug metadata breaking cljs compilation
fixes core.match usage, which can include a
clojure.lang.IPersistentVector class tag in certain patterns. This class
object ended up in the debug metadata and cljs could not serialize it.

WARN: keeping arbitrary metadata is dangerous as there might be other
non-serializable values. We fixed this 1 edge case and left the rest
as-is for now, but a more general solution might be neede in the future.