send heartbeats from server
Before we added HTTPKit support the server sent PING frames to the
client. The client responded with a PONG. If the connection was idle the
server closed it.

HTTPKit doesn't have websocket PING support [1]. To support and unify with
Jetty we changed heartbeats - now the client sent a special HEARTBEAT
message to the server periodically [2].

This resulted in a regression. Chrome throttles a hidden tab's timers
after 5 minutes [3]. They tick in a batch once a minute. We set the max
allowed idle time for a connection to 59 seconds since we deploy to
fly.io which kills every TCP connection idle for over 1 minute [4]. The
result - in Chrome the connection dies and reconnects in a loop.

This commit overcomes all of the above limitations. The server takes
back the initiative and sends a HEARTBEAT message to every client. The
client sends a HEARTBEAT message back to the server. This way

- we don't rely on websocket PINGs (since HTTPKit doesn't have them)
- we don't get throttled in Chrome (since we now rely on server timers)
- we fit into Fly's 1 minute limit

[1] https://github.com/http-kit/http-kit/issues/350
[2] https://github.com/hyperfiddle/electric/commit/6008d18049560bacda5b8e32c407ddf175add852
[3] https://developer.chrome.com/blog/timer-throttling-in-chrome-88/
[4] https://community.fly.io/t/app-connection-timeout/5369