Optimize electric-dom2 macroexpansion size
`dom/props` and `dom/style` will now only touch CSS properties that changed,
because it is the idiomatic thing to do. At the cost of introducing one e/for
per `dom/style`
Before, we were setting all CSS properties for a node when any of them changed. It was
not a perf problem as browsers won’t trigger a repaint/layout if a property value has not
changed.