Optimize call js globals with the correct `this` context
Previous impl was emitting an IIFE the Google Closure Compiler could not inline.
This change moves the `this` binding from compile time to runtime. The cost is
lower, because the emitted code is smaller and the `this` binding happens on
program initialization, before electric runs.

Example of calling a js global with the wrong `this` context:
`(apply js/alert ["hello"])`
would translate to alert.call(null, ["hello"]), whereas the correct form should
be `alert.call(window, ["hello"])`