Fix js field access generating invalid js fn argument symbol
Electric compiled `(.-field js/three.nesting.level)` to the invalid:
`function($three$nesting.level){ … }`. It now produces `function($three$nesting_level){…}`.

`(fn [a.b])` is invalid in clj and cljs because `a.b` is an illegal field name.
Cljs compiles it to `function(a.b){}` which is also invalid javascript.