compiler: fix cljs var resolving
This commit removes some strictness from the compiler's symbol analysis
and allows unknown symbols to pass through optimistically as vars.
Before we'd not compile and produce a helpful error message. Now the
error message might be less helpful, depending on what happens during
cljs compilation. It might result in an unfriendly runtime error.

The long term goal is to write our own, pure analyzer. This is a short
term fix to unblock DE.

Note this is broken in current v2 IC version too, so this is strictly an
improvement.