Inform users of misaligned client/server DAGs
This was a chaotic situation where both users and us would get confused.
This got us so confused it triggered multiple meetings.

Solution: detect the situation and orient the user about:
- what is happening,
- what to do if this happens in prod.

We also kill the ws connection and inform the electric client to NOT attempt to
reconnect to avoid a broken program reconnection loop.