compiler: lift fn args to nodes when site differs
Also changed ap-collapse logic:

In some cases (inlining nodes) the collapse logic came before and
didn't trigger. Had to refactor a lot of code.

We thought we can rewrite

  (r/ap (r/pure inc) (r/pure 1))

to

  (r/pure (inc 1))

But in the general case that's not true. If we substitute `inc` for
`prn` we get from

  (r/ap (r/pure prn) (r/pure 1))

to

  (r/pure (prn 1))

which will run eagerly, whereas the previous version only runs when
sampled.

An always correct rewrite is

  (r/ap (r/pure (fn [] (prn 1))))