[runtime] lazy drain
We experimented with eager drain. In `(do (a) (b))` we don't care about
the return value of `(a)` and with eager drain we can discard the return
value without notifying through the graph.

The problem is it introduces inconsistencies, e.g. in

    (when-let [x (e/watch !x)]
      (prn x)
      (do-something-with x))

the `(prn x)` form will run even when `x` turns falsy, because of its
over-eagerness.

This commit switches to a lazy drain implementation to fix the
inconsistencies at the cost of additional graph churn.