[lang] clean exceptions
Today when an effect throws we get a stack trace. This stack trace is
nothing like our code. Electric runs asynchronously, damaging the stack
trace.

With this change

- we track file/line info of the effects
- we catch exceptions
- we throw an exception containing only the message and file/line info

This improves debugging considerably.

In v2 we had a complete virtual stack trace tracking user code through
the signal chain. This commit doesn't build that. It's future work.