[LANG] more line/col metadata forwarding
There's a problem with the current approach. As we recurse on a hash map
the map kv entries cannot carry metadata, and at this point we lose
track of line/col metadata information.

There will be an upcoming fix that will store the line/col metadata in
the environment map so we can always carry the information over, even in
presence of non-meta objects along the way.