[electric] fix hot code reload when editing runtime3
The runtime calls `slot-port` during teardown but that happens after
shadow-cljs recompiles the touched runtime file, causing a "cannot cast
Slot to Slot" exception. We turn it into a protocol to overcome this
ordering issue. Ideally we'd tear down before shadow-cljs reloads but
that is not obvious how to do, shadow's `:compile-prepare` hook runs
after the clj (JVM) namespace reload.