[forms] Input and Input! handle maxlength again, use in demos
This commit:

https://github.com/hyperfiddle/electric/commit/c968305ed5f83a1aaa2c750d94fd20d6a72dd5f7

introduced 2 changes in `Input` and `Input!`:
1. it removed the default 100 maxlength
2. it removed maxlength handling altogether

1 was harmful, users complained, it wasn't generic. 2 was useful,
because DOM inputs can have a larger value than the maxlength property
allows them to. The input honors the limit when typing but doesn't
enforce it programmatically.

We implemented 1 to guard inputs that end up stored on the server.
Dropping it we lost this guard.

This commit reinstates 2 in `Input` and `Input!` without reinstating 1.
1 is implemented in userland where needed by passing in `:maxlength 100`
manually. We expect to improve this in the future by providing better,
sanitized input components.